// src/app/folha/tipoafastamento/models/tipoafastamento.ts
export class TipoAfastamento{
    id: string;
    empresaId: number;
    grupoEmpresaId: number;
    codigo: string;
    sigla: string;
    descricao: string;
    tipoDeAfastamento: string;
    remuneracao: string;
    codigoESocial: string;
    tipoSEFIP: string;
    codigoRais: string;
    excluido: string;
}